requirements.txt
Flask
Flask_SQLAlchemy
Flask-Mail
qrcode
gunicorn
python-dotenv
Flask-WTF
reportlab
